Update index.html
Since the link in a sentence was not working I decided to change it to "Click this sentence to visit the changelog."
The version number will change once I find things that need to be fixed or removed.
The words "Checking For Problems" , "Correcting Typos" , "Fixing Glitches" , "Fixing Errors" , "Patching Glitches" , "Patching Errors" , "In Beta Stage: Some Features May Not Work!" and "Official Version" that are next to the number indicate what is being changed. More words will be used when different problems occur.